refactor: share searchable-option and computed-field resolution
Extract FormDefinitions searchOptions()/resolveFields() into a
ResolvesFormFields trait keyed on a formFields() source, and fold the action
validation trait into InteractsWithActionForm which reuses it. No behavior
change; sets up actions to search/resolve against their embedded form.
