+++ src/Actions/ActionDefinition.php
@@ -6,13 +6,13 @@ namespace Lattice\Lattice\Actions;
 
 use Illuminate\Http\Request;
 use Lattice\Lattice\Actions\Components\Action;
-use Lattice\Lattice\Actions\Concerns\ValidatesActionForm;
+use Lattice\Lattice\Actions\Concerns\InteractsWithActionForm;
 use Lattice\Lattice\Actions\Contracts\ProvidesAction;
 use Lattice\Lattice\Core\Definition;
 
 abstract class ActionDefinition extends Definition implements ProvidesAction
 {
-    use ValidatesActionForm;
+    use InteractsWithActionForm;
 
     abstract public function definition(Action $action): Action;
