+++ src/Actions/BulkActionDefinition.php
@@ -7,13 +7,13 @@ namespace Lattice\Lattice\Actions;
 use Illuminate\Http\Request;
 use Illuminate\Support\Collection;
 use Lattice\Lattice\Actions\Components\Action;
-use Lattice\Lattice\Actions\Concerns\ValidatesActionForm;
+use Lattice\Lattice\Actions\Concerns\InteractsWithActionForm;
 use Lattice\Lattice\Actions\Contracts\ProvidesBulkAction;
 use Lattice\Lattice\Core\Definition;
 
 abstract class BulkActionDefinition extends Definition implements ProvidesBulkAction
 {
-    use ValidatesActionForm;
+    use InteractsWithActionForm;
 
     abstract public function definition(Action $action): Action;
